Fix decoration checking
The old design was flawed because the `filter` was lazy, which
lead to the decoration check being run *before* and not after
as intended.

To be on the safe side, we now check both before and after,
which also helps avoid some work if the decoration isn't
present in the first place.

The `collect` isn't really an issue because most of the time
it will be empty anyway.