Simplify decoration check
Now that the screen colors are kept in a "screenshot", the decoration
won't change between the decoration check and the life/es/mana checks
so there is no need to check for decoration after again.

This removes the need for "special" checks, which means logout key
will only work while the game is open and decorations are in place,
but that's fine, because the key can now be pressed out of the game.