Remove ScreenChecker
It's slower, the results won't be as accurate as reading from memory,
and it does not work if there are other effects on screen such as delve
or a value is covered by the decoration around the life or mana pool.

While it's probably safer in that we're not reading values from memory,
it's quite limiting (and performing automatic actions on certain events,
even if checking the screen automatically, is already bannable anyway).