Don't reopen the process every time it is checked
This should be more efficient resource-wise, although now we're giving
PoE more time to check if someone is hooked to it (if that's possible,
at all). However it has always had a chance to do so, even if for a
shorter amount of time, so this should not be a concern.

This does mean the program now won't start unless PoE is running
beforehand.